requirements.new.txt
Flask>=0.9
Flask-Mail>=0.7.4
Flask-SQLAlchemy>=0.16
Jinja2>=2.6
SQLAlchemy>=0.7.9
Werkzeug>=0.8.3
boto>=2.6.0
simplejson>=2.3.2
python-dateutil>=2.1
sh>=1.07
flask-openid
beautifulsoup4
